pyproject.toml
[project]
name = "bitebase-intelligence"
version = "0.1.0"
description = "BiteBase Intelligence - Agentic Workflow Platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "langflow>=1.1.0",
    "langchain>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv",
    "sqlalchemy",
    "pandas",
    "numpy",
    "matplotlib",
    "openai",
    "requests",
    "beautifulsoup4",
]
